backend/requirements.txt
fastapi==0.111.1
uvicorn==0.29.0
pydantic==2.11.3
firebase-admin==6.5.0
google-cloud-language==2.13.4
google-generativeai==0.8.5
faiss-cpu==1.10.0
sentence-transformers==4.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.27.2
starlette==0.37.2